requirements.txt
bilibili-api-python
pandas
openpyxl
fastapi
uvicorn[standard]
